fix présentation images

+++ exile/views.py
@@ -3007,7 +3007,7 @@ def planets(request):
                     planet['normal_hydrocarbon'] = True
                 energy_level = getpercent(re[31], re[14], 10)
                 planet['normal_energy'] = True
-                planet['credits'] = re[37] + (re[38] / 2) # - (re["upkeep"] / 24)
+                planet['credits'] = int(re[37] + (re[38] / 2)) # - (re["upkeep"] / 24)
                 if planet['credits'] < 0:
                     planet['credits_minus'] = True
                 else:
